Refactored into reusable util
