+++ x-pack/test/security_solution_api_integration/test_suites/detections_response/rules_management/prebuilt_rules/management/trial_license_complete_tier/index.ts
@@ -22,7 +22,7 @@ export default ({ loadTestFile }: FtrProviderContext): void => {
     loadTestFile(require.resolve('./upgrade_review_prebuilt_rules.rule_type_fields'));
     loadTestFile(require.resolve('./upgrade_review_prebuilt_rules.number_fields'));
     loadTestFile(require.resolve('./upgrade_review_prebuilt_rules.single_line_string_fields'));
-    loadTestFile(require.resolve('./MOD_upgrade_review_prebuilt_rules.single_line_string_fields'));
+    loadTestFile(require.resolve('./new'));
     loadTestFile(require.resolve('./upgrade_review_prebuilt_rules.scalar_array_fields'));
     loadTestFile(require.resolve('./upgrade_review_prebuilt_rules.multi_line_string_fields'));
     loadTestFile(require.resolve('./upgrade_review_prebuilt_rules.data_source_fields'));
