+++ x-pack/test/security_solution_api_integration/test_suites/detections_response/rules_management/prebuilt_rules/management/trial_license_complete_tier/upgrade_review_prebuilt_rules.single_line_string_fields.ts
@@ -6,10 +6,14 @@
  */
 import expect from 'expect';
 import {
+  CommonFieldsDiff,
+  ThreatMatchFieldsDiff,
+  NewTermsFieldsDiff,
   ThreeWayDiffConflict,
   ThreeWayDiffOutcome,
   ThreeWayMergeOutcome,
 } from '@kbn/security-solution-plugin/common/api/detection_engine';
+import { SINGLE_LINE_STRING_FIELDS } from '@kbn/security-solution-plugin/server/lib/detection_engine/prebuilt_rules/logic/diff/calculation/calculate_rule_fields_diff';
 import { FtrProviderContext } from '../../../../../../ftr_provider_context';
 import {
   deleteAllTimelines,
@@ -18,10 +22,71 @@ import {
   installPrebuiltRules,
   createPrebuiltRuleAssetSavedObjects,
   reviewPrebuiltRulesToUpgrade,
+  updateRule,
+  fetchRule,
   patchRule,
   createHistoricalPrebuiltRuleAssetSavedObjects,
+  createRuleAssetSavedObjectOfType,
 } from '../../../../utils';
 import { deleteAllRules } from '../../../../../../../common/utils/security_solution';
+
+interface SingleLineStringFieldTestValues {
+  baseValue: string;
+  customValue: string;
+  updatedValue: string;
+}
+
+const SINGLE_LINE_STRING_FIELDS_MAP: Record<
+  SINGLE_LINE_STRING_FIELDS,
+  SingleLineStringFieldTestValues
+> = {
+  name: {
+    baseValue: 'Base Rule Name',
+    customValue: 'Custom Rule Name',
+    updatedValue: 'Updated Rule Name',
+  },
+  severity: {
+    baseValue: 'low',
+    customValue: 'medium',
+    updatedValue: 'high',
+  },
+  threat_indicator_path: {
+    baseValue: 'indicator.file.hash.base',
+    customValue: 'indicator.ip.current',
+    updatedValue: 'indicator.domain.target',
+  },
+  history_window_start: {
+    baseValue: 'now-10000m',
+    customValue: 'now-20000m',
+    updatedValue: 'now-30000m',
+  },
+};
+
+const RULE_TYPE_FIELD_MAPPING = {
+  query: ['name', 'severity'],
+  threat_match: ['name', 'severity', 'threat_indicator_path'],
+  new_terms: ['name', 'severity', 'history_window_start'],
+} as const;
+
+type RuleTypeToFields = typeof RULE_TYPE_FIELD_MAPPING;
+
+const getValidFieldsForRuleType = <T extends keyof RuleTypeToFields>(
+  ruleType: T
+): { [K in RuleTypeToFields[T][number]]: SingleLineStringFieldTestValues } => {
+  return (
+    Object.entries(SINGLE_LINE_STRING_FIELDS_MAP) as Array<
+      [SINGLE_LINE_STRING_FIELDS, SingleLineStringFieldTestValues]
+    >
+  )
+    .filter(([field]) =>
+      (RULE_TYPE_FIELD_MAPPING[ruleType] as readonly SINGLE_LINE_STRING_FIELDS[]).includes(field)
+    )
+    .reduce((acc, [field, values]) => ({ ...acc, [field]: values }), {}) as {
+    [K in RuleTypeToFields[T][number]]: SingleLineStringFieldTestValues;
+  };
+};
+
+type DiffFieldsForType<T> = CommonFieldsDiff & T;
 
 export default ({ getService }: FtrProviderContext): void => {
   const es = getService('es');
@@ -29,344 +94,992 @@ export default ({ getService }: FtrProviderContext): void => {
   const log = getService('log');
 
   describe('@ess @serverless @skipInServerlessMKI review prebuilt rules updates from package with mock rule assets', () => {
-    beforeEach(async () => {
-      await deleteAllRules(supertest, log);
-      await deleteAllTimelines(es, log);
-      await deleteAllPrebuiltRuleAssets(es, log);
+    describe('Query rule single line string fields', () => {
+      beforeEach(async () => {
+        await deleteAllRules(supertest, log);
+        await deleteAllTimelines(es, log);
+        await deleteAllPrebuiltRuleAssets(es, log);
+      });
+      const queryRuleFields = getValidFieldsForRuleType('query');
+      (
+        Object.entries(queryRuleFields) as Array<
+          [
+            Extract<SINGLE_LINE_STRING_FIELDS, keyof typeof queryRuleFields>,
+            SingleLineStringFieldTestValues
+          ]
+        >
+      ).map(([field, testValues]) => {
+        const { baseValue, customValue, updatedValue } = testValues;
+
+        describe(`testing field: ${field}`, () => {
+          const getRuleAssetSavedObjects = () => [
+            createRuleAssetSavedObject({
+              rule_id: 'rule-1',
+              version: 1,
+              [field]: baseValue,
+            }),
+          ];
+
+          describe("when rule field doesn't have an update and has no custom value - scenario AAA", () => {
+            it('should not show in the upgrade/_review API response', async () => {
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+              await installPrebuiltRules(es, supertest);
+
+              const updatedRuleAssetSavedObjects = [
+                createRuleAssetSavedObject({
+                  rule_id: 'rule-1',
+                  [field]: baseValue,
+                  version: 2,
+                }),
+              ];
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+              const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+              expect(
+                (reviewResponse.rules[0].diff.fields as DiffFieldsForType<ThreatMatchFieldsDiff>)[
+                  field
+                ]
+              ).toBeUndefined();
+
+              expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
+              expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
+              expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+              expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
+              expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+            });
+          });
+
+          describe("when rule field doesn't have an update but has a custom value - scenario ABA", () => {
+            it('should show in the upgrade/_review API response', async () => {
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+              await installPrebuiltRules(es, supertest);
+
+              await patchRule(supertest, log, {
+                rule_id: 'rule-1',
+                [field]: customValue,
+              });
+
+              const updatedRuleAssetSavedObjects = [
+                createRuleAssetSavedObject({
+                  rule_id: 'rule-1',
+                  [field]: baseValue,
+                  version: 2,
+                }),
+              ];
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+              const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+              expect(
+                (reviewResponse.rules[0].diff.fields as DiffFieldsForType<ThreatMatchFieldsDiff>)[
+                  field
+                ]
+              ).toEqual({
+                base_version: baseValue,
+                current_version: customValue,
+                target_version: baseValue,
+                merged_version: customValue,
+                diff_outcome: ThreeWayDiffOutcome.CustomizedValueNoUpdate,
+                merge_outcome: ThreeWayMergeOutcome.Current,
+                conflict: ThreeWayDiffConflict.NONE,
+                has_update: false,
+                has_base_version: true,
+              });
+
+              expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
+              expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
+              expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+              expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
+              expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+            });
+          });
+
+          describe('when rule field has an update but does not have a custom value - scenario AAB', () => {
+            it('should show in the upgrade/_review API response', async () => {
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+              await installPrebuiltRules(es, supertest);
+
+              const updatedRuleAssetSavedObjects = [
+                createRuleAssetSavedObject({
+                  rule_id: 'rule-1',
+                  version: 2,
+                  [field]: updatedValue,
+                }),
+              ];
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+              const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+              expect(
+                (reviewResponse.rules[0].diff.fields as DiffFieldsForType<ThreatMatchFieldsDiff>)[
+                  field
+                ]
+              ).toEqual({
+                base_version: baseValue,
+                current_version: baseValue,
+                target_version: updatedValue,
+                merged_version: updatedValue,
+                diff_outcome: ThreeWayDiffOutcome.StockValueCanUpdate,
+                merge_outcome: ThreeWayMergeOutcome.Target,
+                conflict: ThreeWayDiffConflict.NONE,
+                has_update: true,
+                has_base_version: true,
+              });
+
+              expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(2);
+              expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
+              expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+              expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
+              expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+            });
+          });
+
+          describe('when rule field has an update and a custom value that are the same - scenario ABB', () => {
+            it('should show in the upgrade/_review API response', async () => {
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+              await installPrebuiltRules(es, supertest);
+
+              await patchRule(supertest, log, {
+                rule_id: 'rule-1',
+                [field]: customValue,
+              });
+
+              const updatedRuleAssetSavedObjects = [
+                createRuleAssetSavedObject({
+                  rule_id: 'rule-1',
+                  version: 2,
+                  [field]: customValue,
+                }),
+              ];
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+              const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+              expect(
+                (reviewResponse.rules[0].diff.fields as DiffFieldsForType<ThreatMatchFieldsDiff>)[
+                  field
+                ]
+              ).toEqual({
+                base_version: baseValue,
+                current_version: customValue,
+                target_version: customValue,
+                merged_version: customValue,
+                diff_outcome: ThreeWayDiffOutcome.CustomizedValueSameUpdate,
+                merge_outcome: ThreeWayMergeOutcome.Current,
+                conflict: ThreeWayDiffConflict.NONE,
+                has_update: false,
+                has_base_version: true,
+              });
+
+              expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
+              expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
+              expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+              expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
+              expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+            });
+          });
+
+          describe('when rule field has an update and a custom value that are different - scenario ABC', () => {
+            it('should show in the upgrade/_review API response', async () => {
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+              await installPrebuiltRules(es, supertest);
+
+              await patchRule(supertest, log, {
+                rule_id: 'rule-1',
+                [field]: customValue,
+              });
+
+              const updatedRuleAssetSavedObjects = [
+                createRuleAssetSavedObject({
+                  rule_id: 'rule-1',
+                  version: 2,
+                  [field]: updatedValue,
+                }),
+              ];
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+              const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+              expect(
+                (reviewResponse.rules[0].diff.fields as DiffFieldsForType<ThreatMatchFieldsDiff>)[
+                  field
+                ]
+              ).toEqual({
+                base_version: baseValue,
+                current_version: customValue,
+                target_version: updatedValue,
+                merged_version: customValue,
+                diff_outcome: ThreeWayDiffOutcome.CustomizedValueCanUpdate,
+                merge_outcome: ThreeWayMergeOutcome.Current,
+                conflict: ThreeWayDiffConflict.NON_SOLVABLE,
+                has_update: true,
+                has_base_version: true,
+              });
+
+              expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(2);
+              expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(1);
+              expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(1);
+
+              expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_conflicts).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(1);
+            });
+          });
+
+          describe('when rule base version does not exist', () => {
+            describe('when rule field has an update and a custom value that are the same - scenario -AA', () => {
+              it('should not show in the upgrade/_review API response', async () => {
+                await createPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+                await installPrebuiltRules(es, supertest);
+
+                await deleteAllPrebuiltRuleAssets(es, log);
+
+                const updatedRuleAssetSavedObjects = [
+                  createRuleAssetSavedObject({
+                    rule_id: 'rule-1',
+                    version: 2,
+                    [field]: baseValue,
+                  }),
+                ];
+                await createPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+                const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+                expect(
+                  (reviewResponse.rules[0].diff.fields as DiffFieldsForType<ThreatMatchFieldsDiff>)[
+                    field
+                  ]
+                ).toBeUndefined();
+
+                expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
+                expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
+                expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+                expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+                expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
+                expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+              });
+            });
+
+            describe('when rule field has an update and a custom value that are different - scenario -AB', () => {
+              it('should show in the upgrade/_review API response', async () => {
+                await createPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+                await installPrebuiltRules(es, supertest);
+
+                await deleteAllPrebuiltRuleAssets(es, log);
+
+                await patchRule(supertest, log, {
+                  rule_id: 'rule-1',
+                  [field]: customValue,
+                });
+
+                const updatedRuleAssetSavedObjects = [
+                  createRuleAssetSavedObject({
+                    rule_id: 'rule-1',
+                    version: 2,
+                    [field]: updatedValue,
+                  }),
+                ];
+                await createPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+                const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+                expect(
+                  (reviewResponse.rules[0].diff.fields as DiffFieldsForType<ThreatMatchFieldsDiff>)[
+                    field
+                  ]
+                ).toEqual({
+                  current_version: customValue,
+                  target_version: updatedValue,
+                  merged_version: updatedValue,
+                  diff_outcome: ThreeWayDiffOutcome.MissingBaseCanUpdate,
+                  merge_outcome: ThreeWayMergeOutcome.Target,
+                  conflict: ThreeWayDiffConflict.SOLVABLE,
+                  has_update: true,
+                  has_base_version: false,
+                });
+
+                expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(2);
+                expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(1);
+                expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+                expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+                expect(reviewResponse.stats.num_rules_with_conflicts).toBe(1);
+                expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+              });
+            });
+          });
+        });
+      });
     });
 
-    describe(`single line string fields`, () => {
-      const getRuleAssetSavedObjects = () => [
-        createRuleAssetSavedObject({ rule_id: 'rule-1', version: 1, name: 'A' }),
-      ];
-
-      describe("when rule field doesn't have an update and has no custom value - scenario AAA", () => {
-        it('should not show in the upgrade/_review API response', async () => {
-          // Install base prebuilt detection rule
-          await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
-          await installPrebuiltRules(es, supertest);
-
-          // Increment the version of the installed rule, do NOT update the related single line string field, and create the new rule assets
-          const updatedRuleAssetSavedObjects = [
-            createRuleAssetSavedObject({
-              rule_id: 'rule-1',
-              name: 'A',
-              version: 2,
-            }),
-          ];
-          await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
-
-          // Call the upgrade review prebuilt rules endpoint and check that there is 1 rule eligible for update
-          // but single line string field (name) is NOT returned
-          const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
-          expect(reviewResponse.rules[0].diff.fields.name).toBeUndefined();
-
-          expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
-          expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
-          expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
-
-          expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
-          expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
-          expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
-        });
-
-        it('should trim all whitespace before version comparison', async () => {
-          // Install base prebuilt detection rule
-          await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
-          await installPrebuiltRules(es, supertest);
-
-          // Customize a single line string field on the installed rule
-          await patchRule(supertest, log, {
-            rule_id: 'rule-1',
-            name: 'A\n',
-          });
-
-          // Increment the version of the installed rule, do NOT update the related single line string field, and create the new rule assets
-          const updatedRuleAssetSavedObjects = [
-            createRuleAssetSavedObject({
-              rule_id: 'rule-1',
-              name: '\nA',
-              version: 2,
-            }),
-          ];
-          await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
-
-          // Call the upgrade review prebuilt rules endpoint and check that there is 1 rule eligible for update
-          // but single line string field (name) is NOT returned
-          const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
-          expect(reviewResponse.rules[0].diff.fields.name).toBeUndefined();
-
-          expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
-          expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
-          expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
-
-          expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
-          expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
-          expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
-        });
+    describe('Threat match rule single line string fields', () => {
+      const threatMatchRuleFields = getValidFieldsForRuleType('threat_match');
+      beforeEach(async () => {
+        await deleteAllRules(supertest, log);
+        await deleteAllTimelines(es, log);
+        await deleteAllPrebuiltRuleAssets(es, log);
       });
 
-      describe("when rule field doesn't have an update but has a custom value - scenario ABA", () => {
-        it('should show in the upgrade/_review API response', async () => {
-          // Install base prebuilt detection rule
-          await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
-          await installPrebuiltRules(es, supertest);
+      (
+        Object.entries(threatMatchRuleFields) as Array<
+          [
+            Extract<SINGLE_LINE_STRING_FIELDS, keyof typeof threatMatchRuleFields>,
+            SingleLineStringFieldTestValues
+          ]
+        >
+      ).map(([field, testValues]) => {
+        const { baseValue, customValue, updatedValue } = testValues;
 
-          // Customize a single line string field on the installed rule
-          await patchRule(supertest, log, {
-            rule_id: 'rule-1',
-            name: 'B',
-          });
-
-          // Increment the version of the installed rule, do NOT update the related single line string field, and create the new rule assets
-          const updatedRuleAssetSavedObjects = [
-            createRuleAssetSavedObject({
+        describe(`testing field: ${field}`, () => {
+          const getRuleAssetSavedObjects = () => [
+            createRuleAssetSavedObjectOfType('threat_match', {
               rule_id: 'rule-1',
-              name: 'A',
-              version: 2,
+              version: 1,
+              [field]: baseValue,
             }),
           ];
-          await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
 
-          // Call the upgrade review prebuilt rules endpoint and check that single line string diff field
-          // is returned but field does not have an update, and the merge outcome is "Current"
-          const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
-          expect(reviewResponse.rules[0].diff.fields.name).toEqual({
-            base_version: 'A',
-            current_version: 'B',
-            target_version: 'A',
-            merged_version: 'B',
-            diff_outcome: ThreeWayDiffOutcome.CustomizedValueNoUpdate,
-            merge_outcome: ThreeWayMergeOutcome.Current,
-            conflict: ThreeWayDiffConflict.NONE,
-            has_update: false,
-            has_base_version: true,
+          describe("when rule field doesn't have an update and has no custom value - scenario AAA", () => {
+            it('should not show in the upgrade/_review API response', async () => {
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+              await installPrebuiltRules(es, supertest);
+
+              const updatedRuleAssetSavedObjects = [
+                createRuleAssetSavedObjectOfType('threat_match', {
+                  rule_id: 'rule-1',
+                  [field]: baseValue,
+                  version: 2,
+                }),
+              ];
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+              const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+
+              expect(
+                (reviewResponse.rules[0].diff.fields as DiffFieldsForType<ThreatMatchFieldsDiff>)[
+                  field
+                ]
+              ).toBeUndefined();
+
+              expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
+              expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
+              expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+              expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
+              expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+            });
           });
 
-          expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
-          expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
-          expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+          describe("when rule field doesn't have an update but has a custom value - scenario ABA", () => {
+            it('should show in the upgrade/_review API response', async () => {
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
 
-          expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
-          expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
-          expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
-        });
-      });
+              await installPrebuiltRules(es, supertest);
 
-      describe('when rule field has an update but does not have a custom value - scenario AAB', () => {
-        it('should show in the upgrade/_review API response', async () => {
-          // Install base prebuilt detection rule
-          await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
-          await installPrebuiltRules(es, supertest);
+              const rule = await fetchRule(supertest, { ruleId: 'rule-1' });
 
-          // Increment the version of the installed rule, update a single line string field, and create the new rule assets
-          const updatedRuleAssetSavedObjects = [
-            createRuleAssetSavedObject({
-              rule_id: 'rule-1',
-              version: 2,
-              name: 'B',
-            }),
-          ];
-          await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+              await updateRule(supertest, {
+                ...rule,
+                id: undefined,
+                [field]: customValue,
+              });
 
-          // Call the upgrade review prebuilt rules endpoint and check that one rule is eligible for update
-          const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
-          expect(reviewResponse.rules[0].diff.fields.name).toEqual({
-            base_version: 'A',
-            current_version: 'A',
-            target_version: 'B',
-            merged_version: 'B',
-            diff_outcome: ThreeWayDiffOutcome.StockValueCanUpdate,
-            merge_outcome: ThreeWayMergeOutcome.Target,
-            conflict: ThreeWayDiffConflict.NONE,
-            has_update: true,
-            has_base_version: true,
+              const updatedRuleAssetSavedObjects = [
+                createRuleAssetSavedObjectOfType('threat_match', {
+                  rule_id: 'rule-1',
+                  [field]: baseValue,
+                  version: 2,
+                }),
+              ];
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+              const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+              expect(
+                (
+                  reviewResponse.rules[0].diff
+                    .fields as DiffFieldsForType<ThreatMatchFieldsDiff> as DiffFieldsForType<ThreatMatchFieldsDiff>
+                )[field]
+              ).toEqual({
+                base_version: baseValue,
+                current_version: customValue,
+                target_version: baseValue,
+                merged_version: customValue,
+                diff_outcome: ThreeWayDiffOutcome.CustomizedValueNoUpdate,
+                merge_outcome: ThreeWayMergeOutcome.Current,
+                conflict: ThreeWayDiffConflict.NONE,
+                has_update: false,
+                has_base_version: true,
+              });
+
+              // console.log('HERE', JSON.stringify(reviewResponse, null, 2));
+
+              expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
+              expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
+              expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+              expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
+              expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+            });
           });
 
-          expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(2);
-          expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
-          expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+          describe('when rule field has an update but does not have a custom value - scenario AAB', () => {
+            it('should show in the upgrade/_review API response', async () => {
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+              await installPrebuiltRules(es, supertest);
 
-          expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
-          expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
-          expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
-        });
-      });
+              const updatedRuleAssetSavedObjects = [
+                createRuleAssetSavedObjectOfType('threat_match', {
+                  rule_id: 'rule-1',
+                  version: 2,
+                  [field]: updatedValue,
+                }),
+              ];
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
 
-      describe('when rule field has an update and a custom value that are the same - scenario ABB', () => {
-        it('should show in the upgrade/_review API response', async () => {
-          // Install base prebuilt detection rule
-          await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
-          await installPrebuiltRules(es, supertest);
+              const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+              expect(
+                (reviewResponse.rules[0].diff.fields as DiffFieldsForType<ThreatMatchFieldsDiff>)[
+                  field
+                ]
+              ).toEqual({
+                base_version: baseValue,
+                current_version: baseValue,
+                target_version: updatedValue,
+                merged_version: updatedValue,
+                diff_outcome: ThreeWayDiffOutcome.StockValueCanUpdate,
+                merge_outcome: ThreeWayMergeOutcome.Target,
+                conflict: ThreeWayDiffConflict.NONE,
+                has_update: true,
+                has_base_version: true,
+              });
 
-          // Customize a single line string field on the installed rule
-          await patchRule(supertest, log, {
-            rule_id: 'rule-1',
-            name: 'B',
+              expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(2);
+              expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
+              expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+              expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
+              expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+            });
           });
 
-          // Increment the version of the installed rule, update a single line string field, and create the new rule assets
-          const updatedRuleAssetSavedObjects = [
-            createRuleAssetSavedObject({
-              rule_id: 'rule-1',
-              version: 2,
-              name: 'B',
-            }),
-          ];
-          await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+          describe('when rule field has an update and a custom value that are the same - scenario ABB', () => {
+            it('should show in the upgrade/_review API response', async () => {
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+              await installPrebuiltRules(es, supertest);
 
-          // Call the upgrade review prebuilt rules endpoint and check that one rule is eligible for update
-          const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
-          expect(reviewResponse.rules[0].diff.fields.name).toEqual({
-            base_version: 'A',
-            current_version: 'B',
-            target_version: 'B',
-            merged_version: 'B',
-            diff_outcome: ThreeWayDiffOutcome.CustomizedValueSameUpdate,
-            merge_outcome: ThreeWayMergeOutcome.Current,
-            conflict: ThreeWayDiffConflict.NONE,
-            has_update: false,
-            has_base_version: true,
+              const rule = await fetchRule(supertest, { ruleId: 'rule-1' });
+
+              await updateRule(supertest, {
+                ...rule,
+                id: undefined,
+                [field]: customValue,
+              });
+
+              const updatedRuleAssetSavedObjects = [
+                createRuleAssetSavedObjectOfType('threat_match', {
+                  rule_id: 'rule-1',
+                  version: 2,
+                  [field]: customValue,
+                }),
+              ];
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+              const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+              expect(
+                (reviewResponse.rules[0].diff.fields as DiffFieldsForType<ThreatMatchFieldsDiff>)[
+                  field
+                ]
+              ).toEqual({
+                base_version: baseValue,
+                current_version: customValue,
+                target_version: customValue,
+                merged_version: customValue,
+                diff_outcome: ThreeWayDiffOutcome.CustomizedValueSameUpdate,
+                merge_outcome: ThreeWayMergeOutcome.Current,
+                conflict: ThreeWayDiffConflict.NONE,
+                has_update: false,
+                has_base_version: true,
+              });
+
+              expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
+              expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
+              expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+              expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
+              expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+            });
           });
 
-          expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
-          expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
-          expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+          describe('when rule field has an update and a custom value that are different - scenario ABC', () => {
+            it('should show in the upgrade/_review API response', async () => {
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+              await installPrebuiltRules(es, supertest);
 
-          expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
-          expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
-          expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
-        });
-      });
+              const rule = await fetchRule(supertest, { ruleId: 'rule-1' });
 
-      describe('when rule field has an update and a custom value that are different - scenario ABC', () => {
-        it('should show in the upgrade/_review API response', async () => {
-          // Install base prebuilt detection rule
-          await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
-          await installPrebuiltRules(es, supertest);
+              await updateRule(supertest, {
+                ...rule,
+                id: undefined,
+                [field]: customValue,
+              });
 
-          // Customize a single line string field on the installed rule
-          await patchRule(supertest, log, {
-            rule_id: 'rule-1',
-            name: 'B',
+              const updatedRuleAssetSavedObjects = [
+                createRuleAssetSavedObjectOfType('threat_match', {
+                  rule_id: 'rule-1',
+                  version: 2,
+                  [field]: updatedValue,
+                }),
+              ];
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+              const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+              expect(
+                (reviewResponse.rules[0].diff.fields as DiffFieldsForType<ThreatMatchFieldsDiff>)[
+                  field
+                ]
+              ).toEqual({
+                base_version: baseValue,
+                current_version: customValue,
+                target_version: updatedValue,
+                merged_version: customValue,
+                diff_outcome: ThreeWayDiffOutcome.CustomizedValueCanUpdate,
+                merge_outcome: ThreeWayMergeOutcome.Current,
+                conflict: ThreeWayDiffConflict.NON_SOLVABLE,
+                has_update: true,
+                has_base_version: true,
+              });
+
+              expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(2);
+              expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(1);
+              expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(1);
+
+              expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_conflicts).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(1);
+            });
           });
 
-          // Increment the version of the installed rule, update a single line string field, and create the new rule assets
-          const updatedRuleAssetSavedObjects = [
-            createRuleAssetSavedObject({
-              rule_id: 'rule-1',
-              version: 2,
-              name: 'C',
-            }),
-          ];
-          await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+          describe('when rule base version does not exist', () => {
+            describe('when rule field has an update and a custom value that are the same - scenario -AA', () => {
+              it('should not show in the upgrade/_review API response', async () => {
+                await createPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+                await installPrebuiltRules(es, supertest);
 
-          // Call the upgrade review prebuilt rules endpoint and check that one rule is eligible for update
-          // and single line string field update has NON_SOLVABLE conflict, and merged version is CURRENT
-          const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
-          expect(reviewResponse.rules[0].diff.fields.name).toEqual({
-            base_version: 'A',
-            current_version: 'B',
-            target_version: 'C',
-            merged_version: 'B',
-            diff_outcome: ThreeWayDiffOutcome.CustomizedValueCanUpdate,
-            merge_outcome: ThreeWayMergeOutcome.Current,
-            conflict: ThreeWayDiffConflict.NON_SOLVABLE,
-            has_update: true,
-            has_base_version: true,
-          });
+                await deleteAllPrebuiltRuleAssets(es, log);
 
-          expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(2);
-          expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(1);
-          expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(1);
+                const updatedRuleAssetSavedObjects = [
+                  createRuleAssetSavedObjectOfType('threat_match', {
+                    rule_id: 'rule-1',
+                    version: 2,
+                    [field]: baseValue,
+                  }),
+                ];
+                await createPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+                const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+                expect(
+                  (reviewResponse.rules[0].diff.fields as DiffFieldsForType<ThreatMatchFieldsDiff>)[
+                    field
+                  ]
+                ).toBeUndefined();
 
-          expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
-          expect(reviewResponse.stats.num_rules_with_conflicts).toBe(1);
-          expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(1);
-        });
-      });
+                expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
+                expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
+                expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
 
-      describe('when rule base version does not exist', () => {
-        describe('when rule field has an update and a custom value that are the same - scenario -AA', () => {
-          it('should not show in the upgrade/_review API response', async () => {
-            // Install base prebuilt detection rule
-            await createPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
-            await installPrebuiltRules(es, supertest);
-
-            // Clear previous rule assets
-            await deleteAllPrebuiltRuleAssets(es, log);
-
-            // Increment the version of the installed rule, but keep single line string field unchanged
-            const updatedRuleAssetSavedObjects = [
-              createRuleAssetSavedObject({
-                rule_id: 'rule-1',
-                version: 2,
-                name: 'A', // unchanged
-              }),
-            ];
-            await createPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
-
-            // Call the upgrade review prebuilt rules endpoint and check that one rule is eligible for update
-            // but does NOT contain single line string field
-            const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
-            expect(reviewResponse.rules[0].diff.fields.name).toBeUndefined();
-
-            expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
-            expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
-            expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
-
-            expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
-            expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
-            expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
-          });
-        });
-
-        describe('when rule field has an update and a custom value that are different - scenario -AB', () => {
-          it('should show in the upgrade/_review API response', async () => {
-            // Install base prebuilt detection rule
-            await createPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
-            await installPrebuiltRules(es, supertest);
-
-            // Clear previous rule assets
-            await deleteAllPrebuiltRuleAssets(es, log);
-
-            // Customize a single line string field on the installed rule
-            await patchRule(supertest, log, {
-              rule_id: 'rule-1',
-              name: 'B',
+                expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+                expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
+                expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+              });
             });
 
-            // Increment the version of the installed rule, update a single line string field, and create the new rule assets
-            const updatedRuleAssetSavedObjects = [
-              createRuleAssetSavedObject({
-                rule_id: 'rule-1',
-                version: 2,
-                name: 'C',
-              }),
-            ];
-            await createPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+            describe('when rule field has an update and a custom value that are different - scenario -AB', () => {
+              it('should show in the upgrade/_review API response', async () => {
+                await createPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+                await installPrebuiltRules(es, supertest);
 
-            // Call the upgrade review prebuilt rules endpoint and check that one rule is eligible for update
-            // and single line string field update does not have a conflict
-            const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
-            expect(reviewResponse.rules[0].diff.fields.name).toEqual({
-              current_version: 'B',
-              target_version: 'C',
-              merged_version: 'C',
-              diff_outcome: ThreeWayDiffOutcome.MissingBaseCanUpdate,
-              merge_outcome: ThreeWayMergeOutcome.Target,
-              conflict: ThreeWayDiffConflict.SOLVABLE,
-              has_update: true,
-              has_base_version: false,
+                await deleteAllPrebuiltRuleAssets(es, log);
+
+                const rule = await fetchRule(supertest, { ruleId: 'rule-1' });
+
+                await updateRule(supertest, {
+                  ...rule,
+                  id: undefined,
+                  [field]: customValue,
+                });
+
+                const updatedRuleAssetSavedObjects = [
+                  createRuleAssetSavedObjectOfType('threat_match', {
+                    rule_id: 'rule-1',
+                    version: 2,
+                    [field]: updatedValue,
+                  }),
+                ];
+                await createPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+                const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+                expect(
+                  (reviewResponse.rules[0].diff.fields as DiffFieldsForType<ThreatMatchFieldsDiff>)[
+                    field
+                  ]
+                ).toEqual({
+                  current_version: customValue,
+                  target_version: updatedValue,
+                  merged_version: updatedValue,
+                  diff_outcome: ThreeWayDiffOutcome.MissingBaseCanUpdate,
+                  merge_outcome: ThreeWayMergeOutcome.Target,
+                  conflict: ThreeWayDiffConflict.SOLVABLE,
+                  has_update: true,
+                  has_base_version: false,
+                });
+
+                expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(2);
+                expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(1);
+                expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+                expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+                expect(reviewResponse.stats.num_rules_with_conflicts).toBe(1);
+                expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+              });
+            });
+          });
+        });
+      });
+    });
+
+    describe('New terms rule single line string fields', () => {
+      const newTermsRuleFields = getValidFieldsForRuleType('new_terms');
+      beforeEach(async () => {
+        await deleteAllRules(supertest, log);
+        await deleteAllTimelines(es, log);
+        await deleteAllPrebuiltRuleAssets(es, log);
+      });
+
+      (
+        Object.entries(newTermsRuleFields) as Array<
+          [
+            Extract<SINGLE_LINE_STRING_FIELDS, keyof typeof newTermsRuleFields>,
+            SingleLineStringFieldTestValues
+          ]
+        >
+      ).map(([field, testValues]) => {
+        const { baseValue, customValue, updatedValue } = testValues;
+
+        describe(`testing field: ${field}`, () => {
+          const getRuleAssetSavedObjects = () => [
+            createRuleAssetSavedObjectOfType('new_terms', {
+              rule_id: 'rule-1',
+              version: 1,
+              [field]: baseValue,
+            }),
+          ];
+
+          describe("when rule field doesn't have an update and has no custom value - scenario AAA", () => {
+            it('should not show in the upgrade/_review API response', async () => {
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+              await installPrebuiltRules(es, supertest);
+
+              const updatedRuleAssetSavedObjects = [
+                createRuleAssetSavedObjectOfType('new_terms', {
+                  rule_id: 'rule-1',
+                  [field]: baseValue,
+                  version: 2,
+                }),
+              ];
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+              const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+              expect(
+                (reviewResponse.rules[0].diff.fields as DiffFieldsForType<NewTermsFieldsDiff>)[
+                  field
+                ]
+              ).toBeUndefined();
+
+              expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
+              expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
+              expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+              expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
+              expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+            });
+          });
+
+          describe("when rule field doesn't have an update but has a custom value - scenario ABA", () => {
+            it('should show in the upgrade/_review API response', async () => {
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+              await installPrebuiltRules(es, supertest);
+
+              const rule = await fetchRule(supertest, { ruleId: 'rule-1' });
+
+              await updateRule(supertest, {
+                ...rule,
+                id: undefined,
+                [field]: customValue,
+              });
+
+              const updatedRuleAssetSavedObjects = [
+                createRuleAssetSavedObjectOfType('new_terms', {
+                  rule_id: 'rule-1',
+                  [field]: baseValue,
+                  version: 2,
+                }),
+              ];
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+              const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+              expect(
+                (reviewResponse.rules[0].diff.fields as DiffFieldsForType<NewTermsFieldsDiff>)[
+                  field
+                ]
+              ).toEqual({
+                base_version: baseValue,
+                current_version: customValue,
+                target_version: baseValue,
+                merged_version: customValue,
+                diff_outcome: ThreeWayDiffOutcome.CustomizedValueNoUpdate,
+                merge_outcome: ThreeWayMergeOutcome.Current,
+                conflict: ThreeWayDiffConflict.NONE,
+                has_update: false,
+                has_base_version: true,
+              });
+
+              expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
+              expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
+              expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+              expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
+              expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+            });
+          });
+
+          describe('when rule field has an update but does not have a custom value - scenario AAB', () => {
+            it('should show in the upgrade/_review API response', async () => {
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+              await installPrebuiltRules(es, supertest);
+
+              const updatedRuleAssetSavedObjects = [
+                createRuleAssetSavedObjectOfType('new_terms', {
+                  rule_id: 'rule-1',
+                  version: 2,
+                  [field]: updatedValue,
+                }),
+              ];
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+              const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+              expect(
+                (reviewResponse.rules[0].diff.fields as DiffFieldsForType<NewTermsFieldsDiff>)[
+                  field
+                ]
+              ).toEqual({
+                base_version: baseValue,
+                current_version: baseValue,
+                target_version: updatedValue,
+                merged_version: updatedValue,
+                diff_outcome: ThreeWayDiffOutcome.StockValueCanUpdate,
+                merge_outcome: ThreeWayMergeOutcome.Target,
+                conflict: ThreeWayDiffConflict.NONE,
+                has_update: true,
+                has_base_version: true,
+              });
+
+              expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(2);
+              expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
+              expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+              expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
+              expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+            });
+          });
+
+          describe('when rule field has an update and a custom value that are the same - scenario ABB', () => {
+            it('should show in the upgrade/_review API response', async () => {
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+              await installPrebuiltRules(es, supertest);
+
+              const rule = await fetchRule(supertest, { ruleId: 'rule-1' });
+
+              await updateRule(supertest, {
+                ...rule,
+                id: undefined,
+                [field]: customValue,
+              });
+
+              const updatedRuleAssetSavedObjects = [
+                createRuleAssetSavedObjectOfType('new_terms', {
+                  rule_id: 'rule-1',
+                  version: 2,
+                  [field]: customValue,
+                }),
+              ];
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+              const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+              expect(
+                (reviewResponse.rules[0].diff.fields as DiffFieldsForType<NewTermsFieldsDiff>)[
+                  field
+                ]
+              ).toEqual({
+                base_version: baseValue,
+                current_version: customValue,
+                target_version: customValue,
+                merged_version: customValue,
+                diff_outcome: ThreeWayDiffOutcome.CustomizedValueSameUpdate,
+                merge_outcome: ThreeWayMergeOutcome.Current,
+                conflict: ThreeWayDiffConflict.NONE,
+                has_update: false,
+                has_base_version: true,
+              });
+
+              expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
+              expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
+              expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+              expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
+              expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+            });
+          });
+
+          describe('when rule field has an update and a custom value that are different - scenario ABC', () => {
+            it('should show in the upgrade/_review API response', async () => {
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+              await installPrebuiltRules(es, supertest);
+
+              const rule = await fetchRule(supertest, { ruleId: 'rule-1' });
+
+              await updateRule(supertest, {
+                ...rule,
+                id: undefined,
+                [field]: customValue,
+              });
+
+              const updatedRuleAssetSavedObjects = [
+                createRuleAssetSavedObjectOfType('new_terms', {
+                  rule_id: 'rule-1',
+                  version: 2,
+                  [field]: updatedValue,
+                }),
+              ];
+              await createHistoricalPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+              const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+              expect(
+                (reviewResponse.rules[0].diff.fields as DiffFieldsForType<NewTermsFieldsDiff>)[
+                  field
+                ]
+              ).toEqual({
+                base_version: baseValue,
+                current_version: customValue,
+                target_version: updatedValue,
+                merged_version: customValue,
+                diff_outcome: ThreeWayDiffOutcome.CustomizedValueCanUpdate,
+                merge_outcome: ThreeWayMergeOutcome.Current,
+                conflict: ThreeWayDiffConflict.NON_SOLVABLE,
+                has_update: true,
+                has_base_version: true,
+              });
+
+              expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(2);
+              expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(1);
+              expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(1);
+
+              expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_conflicts).toBe(1);
+              expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(1);
+            });
+          });
+
+          describe('when rule base version does not exist', () => {
+            describe('when rule field has an update and a custom value that are the same - scenario -AA', () => {
+              it('should not show in the upgrade/_review API response', async () => {
+                await createPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+                await installPrebuiltRules(es, supertest);
+
+                await deleteAllPrebuiltRuleAssets(es, log);
+
+                const updatedRuleAssetSavedObjects = [
+                  createRuleAssetSavedObjectOfType('new_terms', {
+                    rule_id: 'rule-1',
+                    version: 2,
+                    [field]: baseValue,
+                  }),
+                ];
+                await createPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+                const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+                expect(
+                  (reviewResponse.rules[0].diff.fields as DiffFieldsForType<NewTermsFieldsDiff>)[
+                    field
+                  ]
+                ).toBeUndefined();
+
+                expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(1);
+                expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(0);
+                expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+                expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+                expect(reviewResponse.stats.num_rules_with_conflicts).toBe(0);
+                expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+              });
             });
 
-            expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(2);
-            expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(1); // name is considered as a conflict
-            expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+            describe('when rule field has an update and a custom value that are different - scenario -AB', () => {
+              it('should show in the upgrade/_review API response', async () => {
+                await createPrebuiltRuleAssetSavedObjects(es, getRuleAssetSavedObjects());
+                await installPrebuiltRules(es, supertest);
 
-            expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
-            expect(reviewResponse.stats.num_rules_with_conflicts).toBe(1);
-            expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+                await deleteAllPrebuiltRuleAssets(es, log);
+
+                const rule = await fetchRule(supertest, { ruleId: 'rule-1' });
+
+                await updateRule(supertest, {
+                  ...rule,
+                  id: undefined,
+                  [field]: customValue,
+                });
+
+                const updatedRuleAssetSavedObjects = [
+                  createRuleAssetSavedObjectOfType('new_terms', {
+                    rule_id: 'rule-1',
+                    version: 2,
+                    [field]: updatedValue,
+                  }),
+                ];
+                await createPrebuiltRuleAssetSavedObjects(es, updatedRuleAssetSavedObjects);
+
+                const reviewResponse = await reviewPrebuiltRulesToUpgrade(supertest);
+                expect(
+                  (reviewResponse.rules[0].diff.fields as DiffFieldsForType<NewTermsFieldsDiff>)[
+                    field
+                  ]
+                ).toEqual({
+                  current_version: customValue,
+                  target_version: updatedValue,
+                  merged_version: updatedValue,
+                  diff_outcome: ThreeWayDiffOutcome.MissingBaseCanUpdate,
+                  merge_outcome: ThreeWayMergeOutcome.Target,
+                  conflict: ThreeWayDiffConflict.SOLVABLE,
+                  has_update: true,
+                  has_base_version: false,
+                });
+
+                expect(reviewResponse.rules[0].diff.num_fields_with_updates).toBe(2);
+                expect(reviewResponse.rules[0].diff.num_fields_with_conflicts).toBe(1);
+                expect(reviewResponse.rules[0].diff.num_fields_with_non_solvable_conflicts).toBe(0);
+
+                expect(reviewResponse.stats.num_rules_to_upgrade_total).toBe(1);
+                expect(reviewResponse.stats.num_rules_with_conflicts).toBe(1);
+                expect(reviewResponse.stats.num_rules_with_non_solvable_conflicts).toBe(0);
+              });
+            });
           });
         });
       });
